리뷰 갯수에 따른 Score 소수 첫째 자리까지 계산 기능 구현

+++ src/main/java/mjc/ramenlog/dto/RestaurantResponseDto.java
@@ -21,7 +21,7 @@ public class RestaurantResponseDto {
     private Address address;
 
     @Schema(description = "식당 평점")
-    private int score;
+    private Double score;
 
 
     public static RestaurantResponseDto from(Restaurant restaurant) {
